refactor(info-panel): restructure birthday and Google meeting components for improved organization and clarity
