--- src/layouts/widgets/wigiPad/info-panel/components/index.ts
@@ -1,3 +1,0 @@
-export { BirthdayItem } from './birthday-item'
-export { NotificationItem } from './notification-item'
-export { GoogleMeetingItem } from './google-meeting-item'
