+++ src/layouts/widgets/wigiPad/info-panel/tabs/google/google-meeting-item.tsx
