+++ src/layouts/widgets/wigiPad/info-panel/tabs/google/google.tab.tsx
@@ -1,0 +1,3 @@
+export function GoogleTab() {
+	return <></>
+}
